feat: added caching of HLS redirect endpoint
