+++ .changeset/fresh-rabbits-bow.md
@@ -1,0 +1,8 @@
+---
+'@livepeer/core': patch
+'@livepeer/core-react': patch
+'@livepeer/core-web': patch
+'@livepeer/react': patch
+---
+
+**Feature:** added shared caching of redirect URL for HLS and WebRTC, for faster HLS boot times.
