+++ packages/core-web/src/media/browser/utils.ts
@@ -1,4 +1,5 @@
 import { Src } from '@livepeer/core/media';
+import fetch from 'cross-fetch';
 
 import { noop } from '../../utils';
 
@@ -77,3 +78,45 @@ export const canPlayMediaNatively = (src: Src): boolean => {
 
   return true;
 };
+
+const DEFAULT_TIMEOUT = 20000;
+
+let cachedRedirectHost: string | null = null;
+
+export async function getRedirectUrl(
+  endpoint: string,
+  abortController: AbortController,
+  timeout?: number,
+) {
+  try {
+    if (cachedRedirectHost) {
+      const inputUrl = new URL(endpoint);
+
+      inputUrl.host = cachedRedirectHost;
+
+      return inputUrl;
+    }
+
+    const id = setTimeout(
+      () => abortController.abort(),
+      timeout ?? DEFAULT_TIMEOUT,
+    );
+
+    const response = await fetch(endpoint, {
+      method: 'GET',
+      signal: abortController.signal,
+    });
+
+    clearTimeout(id);
+
+    const parsedUrl = new URL(response.url);
+
+    if (parsedUrl?.host) {
+      cachedRedirectHost = parsedUrl.host;
+    }
+
+    return parsedUrl;
+  } catch (e) {
+    return null;
+  }
+}
