+++ packages/core-web/src/media/browser/webrtc/whep.ts
@@ -4,9 +4,9 @@ import {
   WebRTCVideoConfig,
   constructClientOffer,
   createPeerConnection,
-  getRedirectUrl,
   negotiateConnectionWithClientOffer,
 } from './shared';
+import { getRedirectUrl } from '../utils';
 
 /**
  * Client that uses WHEP to play back video over WebRTC.
